requirements.txt
bottleneck==1.3.2
Flask==1.1.2
matplotlib==3.3.2
numexpr==2.7.3
numpy==1.19.2
pandas==1.1.2
pdfminer-six==20201018
scipy==1.5.2
tqdm==4.50.2